requirements.txt
mysql-connector-python==9.3.0
python-dotenv==1.1.0